use async/await functions for data fetching
